PROYECTO01 PC - EJB>> 24/02/2014

+++ DesignerRHN/DesignerRHN-ejb/src/java/Entidades/Competenciascargos.java
@@ -6,6 +6,7 @@ package Entidades;
 
 import java.io.Serializable;
 import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Collection;
 import javax.persistence.Basic;
 import javax.persistence.Column;
@@ -39,7 +40,7 @@ public class Competenciascargos implements Serializable {
     @Basic(optional = false)
     @NotNull
     @Column(name = "SECUENCIA")
-    private BigDecimal secuencia;
+    private BigInteger secuencia;
     @Column(name = "PESO")
     private Integer peso;
     @JoinColumn(name = "EVALCOMPETENCIA", referencedColumnName = "SECUENCIA")
@@ -54,15 +55,15 @@ public class Competenciascargos implements Serializable {
     public Competenciascargos() {
     }
 
-    public Competenciascargos(BigDecimal secuencia) {
+    public Competenciascargos(BigInteger secuencia) {
         this.secuencia = secuencia;
     }
 
-    public BigDecimal getSecuencia() {
+    public BigInteger getSecuencia() {
         return secuencia;
     }
 
-    public void setSecuencia(BigDecimal secuencia) {
+    public void setSecuencia(BigInteger secuencia) {
         this.secuencia = secuencia;
     }
